FFWEB-3249: Do not redirect if already on results page
Currently if you enter something in the search box and confirm, you are always redirected to the search results page - even when you are already on the search results page. This fixes that by checking whether we are already on the search results page and only output the factfinder.request.before.search() callback if that is not the case.